requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2017.1.23
config==0.3.9
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11rc1
django-allauth==0.31.0
django-celery-results==1.0.1
emoji==0.4.5
freeze==1.0.10
future==0.16.0
googlemaps==2.4.6
gunicorn==19.7.1
kombu==4.0.2
motionless==1.3.1
numpy==1.12.1
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
pyexcel==0.4.5
pyexcel-io==0.3.3
pyexcel-xls==0.3.1
pyparsing==2.2.0
python-telegram-bot==5.3.0
python3-openid==3.1.0
pytz==2017.2
requests==2.11.1
requests-oauthlib==0.8.0
six==1.10.0
texttable==0.8.8
urllib3==1.20
vine==1.1.3
virtualenv==15.1.0
whitenoise==3.3.0
xlrd==1.0.0
xlwt==1.2.0